requirements.txt
#Wavelink
aiohttp>=3.6.0,<3.7.0
discord.py>=1.2.5
